Fix includes and play with parameters
Some parameter changes to get this to work well in my environment. This
is currently pretty parameter sensitive to the space. I suspect
something can be done about this.

Also changed sensor intrinsics. This probably needs to go in the
parameter file. Added a circular buffer to try and experiment with
better sync between color and depth frames (currently the use of this is
commented out though it is allocated)